+++ test-scripts/run-anchor-tests.sh
@@ -94,6 +94,7 @@ test_files=(
   userAccount.ts
   userDelegate.ts
   userOrderId.ts
+  perpMarketConfig.ts
   # whitelist.ts
 )